Add mappings NSHost, NSProcessInfo, NSWorkspace
Not full mappings mind you, mappings that AXElements
needs. More can be added later.

But really, I would prefer to have a more abstract
mechanism for handling the mappings.

The strategy would be to lazy map methods. We would
only initialize the class and give it a
#method_missing implementation that looks up the
method and if found calls it.

Processing the parameters and return value will
be the hard part. We will have to use things
like method_getTypeEncoding to get the type
encoding, call the right methods to do the wrapping
or unwrapping.

As a final hurdle, memory management will be a bit
of a PITA to figure out, since we should not always
release objects that we get back…hmmm