fix native extension build break when using mismatching OS X and Xcode
Ruby mkmf uses arch versions of current system (e.g. darwin14 for 10.10) but Xcode SDK uses darwin version of that SDK (so 10.11SDK uses darwin15), so it causes compile errors when using Xcode 7 on 10.10. This submission is to fix it by inferring the darwin version from SDK version.

bumping minor version for this is a small non API breaking change.